% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMixing.R
\name{DistMixing}
\alias{DistMixing}
\title{Sampling from two distributions}
\usage{
DistMixing(n, family_1, par_1, family_2, par_2, p = 0.5)
}
\arguments{
\item{n}{sample length}

\item{family_1}{name of the fist distribution}

\item{par_1}{list of parameters 1}

\item{family_2}{name of the second distribution}

\item{par_2}{list of parameters 2}

\item{p}{percentage split of data}
}
\value{
Function returns parameters of used distributions and combinde sample
\item{fam1}{Name of first distribution}
\item{pars1}{Parameters of first distribution}
\item{fam2}{Name of second distribution}
\item{pars2}{Parameters of second distribution}
}
\description{
This function is used to generate samples from two distributions and combining them.
}
\details{
asbnfkjasbfkjasbdf
}
