% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalityTesting.R
\name{NormalityTesting}
\alias{NormalityTesting}
\title{Return data frame with results fasfasfasfasfasf}
\usage{
NormalityTesting(x)
}
\arguments{
\item{x}{sample to test}
}
\value{
Function returns set of outputs (p.value) from normality tests:
\itemize{
    \item Shapiro
    \item JarqueBera
    \item AndersonDarling
    \item KolmogorovSmirnov
    \item PearsonChiSquare
}
}
\description{
Return data frame with results fasfasfasfasfasf
}
\details{
asbnfkjasbfkjasbdf
}
